pyproject.toml
[project]
name = "comparecimento-eleicoes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.43",
    "duckdb>=1.4.4",
    "ipykernel>=7.2.0",
    "pandas>=3.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
